// web/frontend/src/assets/hd-date/types.ts
export enum DateTimeFormat {
  YYYY = 'YYYY',
  yyyy = 'yyyy',
  yy = 'yy',
  MM = 'MM',
  MMM = 'MMM',
  MMMM = 'MMMM',
  Www = 'Www',
  dd = 'dd',
  hh = 'hh',
  HH = 'HH',
  mm = 'mm',
  ss = 'ss',
  tt = 'tt',
}

export enum FullNameDigit {
  Month = 'Month',
  Date = 'Date',
  FullYear = 'FullYear',
  Hours = 'Hours',
  Minutes = 'Minutes',
  Seconds = 'Seconds',
  Milliseconds = 'Milliseconds',
}